Green control keys have returned
(...at request of Lenny)

Now all colour keys work with multipliers.

Green keys do not directly change green, instead increase in green
is a reduction in red and blue and an increase in brightness.
And vice-versa.

This allows the green change to be seen on the graphs, and doesn't
change the meaning of brightness.