File selection dialog when no filename given
Contribution from ItsMeLenny. Tk file dialog appears when
no filename given as argument. This avoids the first use problem
where people run the program and nothing happens.

Possibly in the future this will be replaced by an integrated
browser view but for now it is an improvement on the previous
situation.