First steps to "background" GPU processing using 2nd window/context
The idea is to have a second window with an independant GL context
that can be used for e.g. background MOV export without totally
blocking the main UI.

With GLUT, this can still block because it cannot be multithreaded,
but with GLFW it should be possible to render the background
window from a dedicated render thread.