Add clipping to UI using stencil buffer. Also "Flickable"
Any object can have clipping set to true and its children will
only be drawn on top of the pixels the parent draws to. This
is done using the stencil buffer.

New Flickable element introduced, intended to be used for
lists of stuff larger than the visible area.

Actually flickable is just a draggable for now - it doesn't have
any physics.

A flickable can contains a canvas object as its only child
 which can be dragged in x & y (either can be turned off).

If the canvas has children which respond to input, starting a drag
on one of those is not possible - instead they get the click.

Once a drag has started, you can drag over active children and they
will NOT get any event.