Move export wav file creation into ExportQueue thread
This means you do not need to wait until a file is indexed by the
viewer before it can be added to the export queue.

However, it does mean an MLV file with audio must be indexed first
during export before the rest of the export can complete.

(This is preparing for a soon-to-be-added "export all" option that
adds all files in a directory to the export queue in one go.)