Integrate preprocess shader also with HQ playback mode (VERY SLOW!)
Preprocessing with a shader, then reading back to demosaic with CPU
is very slow. If we had a high quality GPU demosaicer, it would be
much faster to keep everything in GPU ram. But we don't, yet.

It is not yet integrated with the MOV export queue demosaicing,
or the DNG export (as a preprocess step). Those come next.