Fix white balance clicking with preprocessing
Now preprocessed frames record the white balance they were processed with.
Then, if the use clicks to select a white balance, this balance is removed
from the frame so that the correct neautral value is taken.

Also, the rendering logic is updated to allow the white balance to be changed
against a preprocessed & CPU demosaiced frame with just a shader pass.
This allows the balance box handle to be dragged with immediate feedback.