Add balance and in/out marks to persisted values
Reworked the setting/getting of metadata to be be thread safe.

Note: Values that not have been explicitly set in a file currently
float, and inherit the value from the previously viewed file,
until they are explicitly set. This may not be totally desirable,
but it allows many files to be viewed with same white balance, for
instance.