Add framework for 1D->3D->1D LUT
Demosaic shaders now support pipeline of up to 3 LUTs
First 1D LUT is intended later to replace the old
hard-coded mapping functions with a wider range of
included or user supplied curves, including a range
of different log curves.
The 3D LUT is intended for colour style changes
or film emulation.
The final 1D LUT is intended for final gamma curve
changes, if that's not built into the 3D LUT.
All are optional, but later the first 1D LUT will
almost always need to be supplied to avoid
(non-useful) linear mapping.