Initial slow reference Lossless JPEG (92) decoder
The DNG.py script is now capable of partially decoding
e.g. BMPC compressed 12bit DNGs. However it is
a) Slow
b) Not very useful, since it writes the 2 subframes
into seperate PGM files with pairs of rows side by side.
This is how it's encoded in DNG.

This implementation will not be taken into use with
MlRawViewer. But maybe a later one written in C, will.