Add support for directories containing unformatted 16bit raw data files
A file named <something>.rawseq must be in the directory.
It must be a text file with 3 lines: width, height, cfa layout, e.g.

1920
1080
RGGB

CFA layout can currently be "RGGB" or "GBRG"

Open MlRawViewer with the filename of the .rawseq file and it will
attempt to sort and show all the other files in the dir.